refactor: Move navigationMenu to `@/constants` file for better organization
